M9_Lab Activity While Loops
While Statement 
The while statement creates a loop that executes a specified statement as long as the test condition evaluates to true. The condition is evaluated before executing the statement.

condition
An expression is evaluated before each passes through the loop. If this condition evaluates to true, the statement is executed. When the condition evaluates to false, execution continues with the statement after the while loop.